logservice refactor
