+++ src/services/encryption_deprecated/EncryptionServiceBase.ts
@@ -2,7 +2,7 @@ import { Volume } from "../../entities/Volume";
 import { EncryptionService } from "./EncryptionService";
 import PasswordService from "../password/PasswordService";
 
-import log from "../../utils/LogUtil";
+import log from "../LogService";
 import * as ShellHelper from "../../utils/ShellUtil";
 
 export abstract class EncryptionServiceBase implements EncryptionService {
