+++ src/services/password/PasswordService.ts
@@ -1,7 +1,7 @@
 import { ServiceError, ErrorType } from "../ServiceError";
 import { Volume } from "../../entities/Volume";
 import { constants } from "../../utils/constants";
-import log from "../../utils/LogUtil";
+import log from "../LogService";
 import * as keytar from "keytar";
 import PasswordServiceInterface from "./PasswordServiceInterface";
