+++ src/tests/VolumeService.spec.ts
@@ -10,7 +10,7 @@ import * as path from "path";
 
 
 // // disbaling logs for cleaner stdout (is this a good thing???)
-import log from "../utils/LogUtil";
+import log from "../services/LogService";
 log.transports.file.level = false;
 log.transports.console.level = false;
